requirements.txt
opencv-python
numpy
pywebview
streamlit
tk